Simplify statements about authentication of cert information (#243)
- Simplify 3.2.2 to more directly reflect the language used in that
section of the BRs
- Replace sections 3.2.3, 3.2.4, and 3.2.5 with "No applicable", because
Let's Encrypt does not need to perform authentication of individual
identity or validation of authority, and does not include non-verified
subscriber information in certificates

Note that this is the first use of "Not applicable." as full section
contents in this document. This feels more appropriate than "No
stipulation", as we are affirmatively stating that these sections do not
concern our operations, rather than saying simply that we choose not to
describe our operations in these sections.